requirements.txt
alembic==1.4.3
arabic-reshaper==2.1.0
blinker==1.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.2
pandas==1.1.3
pdfkit==0.6.1
Pillow==8.0.1
PyJWT==1.7.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
pytz==2020.1
reportlab==3.5.54
six==1.15.0
SQLAlchemy==1.3.20
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
xhtml2pdf==0.2.5
xlrd==1.2.0
